refactor: centralise adapter session path predicates to adapterPredicates.ts
Move isJetBrainsSessionPath, isCopilotCliSessionPath, and isCopilotChatSessionPath
from their individual adapter files into a single adapterPredicates.ts module.

- isJetBrainsSessionPath was already used across module boundaries by extension.ts
  and usageAnalysis.ts; those imports are updated to adapterPredicates.ts
- isCopilotChatNonSessionFile (renamed from isNonSessionFile) and its
  NON_SESSION_PATTERNS constant are moved alongside isCopilotChatSessionPath
  so the predicate's helper stays co-located with the predicate itself
- copilotChatAdapter.ts imports isCopilotChatNonSessionFile for directory scanning
- Test imports updated to import predicates from adapterPredicates.ts
